+++ packages/material-top-tabs/src/types.tsx
@@ -116,7 +116,9 @@ export type MaterialTopTabNavigationOptions = {
    */
   tabBarIndicator?: (
     props: Omit<
-      Parameters<React.ComponentProps<typeof TabBar>['renderIndicator']>[0],
+      Parameters<
+        NonNullable<React.ComponentProps<typeof TabBar>['renderIndicator']>
+      >[0],
       'navigationState'
     > & { state: TabNavigationState<ParamListBase> }
   ) => React.ReactNode;
@@ -201,6 +203,12 @@ export type MaterialTopTabNavigationOptions = {
   swipeEnabled?: boolean;
 
   /**
+   * Whether to enable animations when switching between tabs by pressing on the tab bar or programmatically.
+   * Switching tab via swipe gesture will still result in an animation.
+   */
+  animationEnabled?: boolean;
+
+  /**
    * Whether this screen should be lazily rendered. When this is set to `true`,
    * the screen will be rendered as it comes into the viewport.
    * By default all screens are rendered to provide a smoother swipe experience.
@@ -251,6 +259,7 @@ export type MaterialTopTabNavigationConfig = Omit<
   | 'renderTabBar'
   | 'renderLazyPlaceholder'
   | 'swipeEnabled'
+  | 'animationEnabled'
   | 'lazy'
   | 'lazyPreloadDistance'
   | 'lazyPlaceholder'